requirements.txt
streamlit
mysql-connector-python
pytesseract
scikit-learn
pandas
Pillow
python-dotenv
bcrypt
faker
reportlab
opencv-python
numpy
extra-streamlit-components
